+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Internal
 * Log the `Content-Type`, `Origin` and `Sec-Fetch-*` header values for each request.
+* Added request completion time logging.
 
 ## [1.6.0] Improved Key-Capsule Expiry Logic & secp265 Support (2025-12-01)